clinical-view-frontend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
ultralytics==8.3.0
opencv-python==4.10.0.84
pillow==11.0.0
numpy==1.26.4
torch==2.5.0
torchvision==0.20.0
